Correçções de Classes DTO e MODELs IMPLEMENTAÇÂOES: UserAuthentication e seu repository, controller e DTO. Criações dos ENDPOINTS MovimentaçãoID, Criação da arquitetura de criptografia com springSecurity

+++ src/main/java/com/dio/santander/bankLine/api/DTO/NovaMovimentacao.java
@@ -15,6 +15,6 @@ public class NovaMovimentacao {
 
     private MovimentacaoTipo tipo;
 
-    private Integer idConta;
+    private Long idConta;
 
 }
