+++ src/main/java/com/dio/santander/bankLine/api/Repository/CorrentistaRepository.java
@@ -3,6 +3,8 @@ package com.dio.santander.bankLine.api.Repository;
 import com.dio.santander.bankLine.api.Model.Correntista;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CorrentistaRepository extends JpaRepository<Correntista, Integer > {
+
+
+public interface CorrentistaRepository extends JpaRepository<Correntista, Long > {
 
 }
