+++ src/main/java/com/dio/santander/bankLine/api/Repository/MovimentacaoRepository.java
@@ -3,5 +3,7 @@ package com.dio.santander.bankLine.api.Repository;
 import com.dio.santander.bankLine.api.Model.Movimentacao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MovimentacaoRepository extends JpaRepository<Movimentacao, Integer> {
+
+
+public interface MovimentacaoRepository extends JpaRepository<Movimentacao, Long> {
 }
